move CudaThreadIndexing to pyccel/cuda
