+++ pyccel/decorators.py
@@ -6,7 +6,7 @@
 """
 This module contains all the provided decorator methods.
 """
-from pyccel.ast.cudaext import cuda_mod
+from pyccel.cuda.cuda_thread_indexing import CudaThreadIndexing
 import warnings
 
 __all__ = (
@@ -131,25 +131,6 @@ def kernel(f):
     KernelAccessor
         A class representing the kernel function.
     """
-    class CudaThreadIndexing:
-        """
-        Class representing the CUDA thread indexing.
-
-        Class representing the CUDA thread indexing.
-        """
-        def __init__(self, block_idx, thread_idx):
-            self._block_idx = block_idx
-            self._thread_idx = thread_idx
-
-        def threadIdx(self, dim):
-            return self._thread_idx
-
-        def blockIdx(self, dim):
-            return self._block_idx
-
-        def blockDim(self, dim):
-            return 0
-
     class KernelAccessor:
         """
         Class representing the kernel function.
@@ -168,7 +149,7 @@ def kernel(f):
                 """
                 for b in range(num_blocks):
                     for t in range(num_threads):
-                        self._f.__globals__['cuda'].CudaThreadIndexing = CudaThreadIndexing(b, t)
+                        self._f.__globals__['cuda'] = CudaThreadIndexing(b, t)
                         self._f(*args, **kwargs)
             return internal_loop
