+++ tests/pyccel/scripts/kernel/hello_kernel.py
@@ -1,6 +1,5 @@
 # pylint: disable=missing-function-docstring, missing-module-docstring
 from pyccel.decorators import kernel
-from pyccel import cuda
 
 @kernel
 def say_hello(its_morning : bool):
@@ -12,6 +11,7 @@ def say_hello(its_morning : bool):
 def f():
     its_morning = True
     say_hello[5,5](its_morning)
+    from pyccel import cuda
     cuda.synchronize()
 
 if __name__ == '__main__':
